// Frontend/src/app/core/interfaces/empleadoCrear.interface.ts
export interface EmpleadoCrear {
    name: string;
    lastName: string;
    dni: string;
    address: string;
    email: string;
    password: string;
    phone: string;
    dateOfBirth: string;
    location: string;
    iban: string;
    idRole: number;
    idPosition: number;
    startDate: string;
    active: boolean
}